requirements.txt
Flask
bootstrap-flask
Flask-SQLAlchemy
Flask-WTF
python-dotenv
requests
wtforms
sqlalchemy
flask-bootstrap